+++ src/Embedders/Embedder.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Rubix\ML\Embedders;
+
+use Rubix\ML\Datasets\Dataset;
+
+interface Embedder
+{
+    /**
+     * Return the data types that this embedder is compatible with.
+     *
+     * @return int[]
+     */
+    public function compatibility() : array;
+
+    /**
+     * Embed a high dimensional dataset into a lower dimensional one.
+     *
+     * @param \Rubix\ML\Datasets\Dataset $dataset
+     * @throws \InvalidArgumentException
+     * @return \Rubix\ML\Datasets\Dataset
+     */
+    public function embed(Dataset $dataset) : Dataset;
+}